requirements.txt
Django==1.9.7
factory_boy==2.7.0
